@@ -57,7 +57,7 @@ disallow_untyped_defs = "True"
 plugins = ['pydantic.mypy']
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
