@@ -5,9 +5,9 @@ build-backend = "pdm.backend"
 [project]
 authors = []
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "langchain-core<1.0.0,>=0.3.70",
+    "langchain-core<1.0.0,>=0.3.75",
     "tokenizers>=0.19.1",
     "huggingface-hub>=0.33.4",
 ]
@@ -55,7 +55,7 @@ langchain-tests = { path = "../../standard-tests", editable = true }
 disallow_untyped_defs = "True"
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 select = [
