@@ -4,8 +4,8 @@ from __future__ import annotations
 
 import logging
 import os
-from collections.abc import AsyncIterator, Awaitable, Iterator
-from typing import Any, Callable, Optional, TypedDict, TypeVar, Union
+from collections.abc import AsyncIterator, Awaitable, Callable, Iterator
+from typing import Any, Literal, Optional, TypedDict, TypeVar, Union
 
 import openai
 from langchain_core.language_models import LanguageModelInput
@@ -16,7 +16,7 @@ from langchain_core.runnables import Runnable
 from langchain_core.utils import from_env, secret_from_env
 from langchain_core.utils.pydantic import is_basemodel_subclass
 from pydantic import BaseModel, Field, SecretStr, model_validator
-from typing_extensions import Literal, Self
+from typing_extensions import Self
 
 from langchain_openai.chat_models.base import BaseChatOpenAI
 
